add InMemoryLockAndTimestampServiceFactory
