+++ atlasdb-config/src/main/java/com/palantir/atlasdb/factory/TransactionManagers.java
@@ -355,8 +355,7 @@ public abstract class TransactionManagers {
 
         Supplier<ManagedTimestampService> managedTimestampSupplier = atlasFactory::getManagedTimestampService;
 
-        LockAndTimestampServices lockAndTimestampServices =
-                LockAndTimestampServiceFactory.createLockAndTimestampServices(
+        LockAndTimestampServices lockAndTimestampServices = new DefaultLockAndTimestampServiceFactory(
                         metricsManager,
                         config(),
                         runtime,
@@ -369,7 +368,8 @@ public abstract class TransactionManagers {
                         reloadingFactory(),
                         timeLockFeedbackBackgroundTask,
                         timelockRequestBatcherProviders(),
-                        schemas());
+                        schemas())
+                .createLockAndTimestampServices();
         adapter.setTimestampService(lockAndTimestampServices.managedTimestampService());
 
         KvsProfilingLogger.setSlowLogThresholdMillis(config().getKvsSlowLogThresholdMillis());
@@ -571,7 +571,7 @@ public abstract class TransactionManagers {
             UserAgent userAgent,
             DialogueClients.ReloadingFactory reloadingFactory) {
         Refreshable<ServerListConfig> serverListConfigSupplier =
-                LockAndTimestampServiceFactory.getServerListConfigSupplierForTimeLock(config, runtimeConfig);
+                DefaultLockAndTimestampServiceFactory.getServerListConfigSupplierForTimeLock(config, runtimeConfig);
 
         BroadcastDialogueClientFactory broadcastDialogueClientFactory = BroadcastDialogueClientFactory.create(
                 reloadingFactory,
@@ -901,7 +901,7 @@ public abstract class TransactionManagers {
             TimestampStoreInvalidator invalidator,
             String userAgent) {
         LockAndTimestampServices lockAndTimestampServices =
-                LockAndTimestampServiceFactory.createRawInstrumentedServices(
+                DefaultLockAndTimestampServiceFactory.createRawInstrumentedServices(
                         metricsManager,
                         config,
                         runtimeConfigSupplier,
